# Add glob support in the finder