Fix #18.

With that update if you specify a path like "/*emp" or "/{a,b,c}*.php", it will be processed nicely.

Example:
```php
<?php
$finder = (new Finder())->in(["/*emp", "/{a,b,c}*.php"]);
```

Does not use the `Iterator\Glob` implementation here because it does not have the GLOB_ONLYDIR filter. It can be replaced with a combination of `Iterator\CallbackFilter` and `Iterator\Glob` but it seems more complex which is not suitable here...