@Hywan 2015-11-13
@shulard Why `GLOB_ONLYDIR` is required?