@Hywan 2015-11-13
@shulard How performances are impacted? The fact we are applying `glob` in all cases makes me unconfortable. Is it possible to detect a glob pattern when using `in` and doing the appropriated computations here?