@shulard 2015-11-13
I used the `GLOB_ONLYDIR` because `glob` list all items by default (files and directories).
The path is used to build a `Iterator\Recursive\Directory` which only handles directories. It's iterator's goal to retrieve all the items that must be listed.