@shulard 2015-11-13
@Hywan maybe we can add a pattern to avoid applying glob on each path.
I though that `glob` default behaviour is to handle that pattern. It always return an array with the matching paths (with dynamic patterns or not). So it's the same as applying a pattern before use it for me :smile:

Also when `glob` will be able to handle more pattern cases, we can use it directly...