@shulard 2015-11-23
Hello,

With my last comment, do you think we really need to apply a `glob` pattern on each path before calling `glob`?