@Hywan 2015-11-24
Can you compare the performances between `GlobIterator` + `FilterIterator` vs. `RegexIterator` vs. `glob`. I reckon `RegexIterator` may be faster than `GlobIterator` and a “`glob2pcre`” algorithm is not very complicated at all.