@shulard 2015-11-24
Hello @Hywan,

Of cours, I can compare performance and I'll submit the bench here. Yes the "`glob2pcre`" algorithm is not very complicated and it allow us to rely only on Iterators...