@shulard 2016-01-12
Hello @Hywan,

I've worked on this PR today and I've commited the benchmark code here: https://github.com/shulard/hoa-file-glob-benchmark

I've used the `Hoa\Bench` lib for timing check and the result is that `GlobIterator`+`CallBackFilter` and `glob`+`GLOB_ONLYDIR` take the same amount of time.

I'm not particularly familiar with code benchmarking, if you have any ideas to improve the check I'm open to add it !

The test was run with a `maxDepth` of 5 on different directories :

* Discover and list 31 files :
  * `glob` take 2ms
  * `GlobIterator` take 2ms

* Discover and list 3708 files :
  * `glob` take 46ms
  * `GlobIterator` take 50ms

* Discover and list 66191 files (ran 8 times) :
  * `glob` take 1087 to 1183ms
  * `GlobIterator` take 1081 to 1193ms


I can't find a way to implement the `RegexIterator` method. If we want to use that, we need to crawl the folder and apply the regex generated so there are some problems :

* We need to find a folder where starting the crawl 
* We need to compute the regex (not complicated here)

This is a lot of useless computing because it require to develop the whole glob process...

What are you thinking about that update ?

Thanks
Stéphane