@shulard 2016-01-13
@Hywan,

I've updated and rebased the code... Now the `in` method contains the glob application.

I've also benched the Finder object with a new test : Try to crawl a path without `glob` pattern inside
The goal of this test is to show `glob` call overhead. The execution time is similar in all the 3 cases :

* Without `glob` pattern computation
* With `glob` use
* With `GlobIterator` use

I've updated my repo to handle that case.