@Jir4 2016-02-17
The PR is good for me, the code is simplier to understand with a simple `glob` use than with a `GlobIterator`use and the bench are similar.