@Hywan 2016-02-17
IRC logs can be interesting: https://botbot.me/freenode/hoaproject/2016-02-17/?msg=60227978&page=1.