@Jir4 2016-02-18
So we have to test if we have to use glob or not. Is there a syntax to detect ?