@Hywan 2016-02-18
https://en.wikipedia.org/wiki/Glob_%28programming%29#Syntax