@shulard 2016-02-19
@Jir4, if you want I can add the syntax detection behaviour in the current code (or you want to work on it ?)...