@shulard 2016-02-19
I've used a basic regex to detect all the components of glob syntax: `/\*|\?|\[([^\]]+)\]|\{([^}]+)\}/`

I've added the braces because in PHP glob it is possible to use them... Maybe we prefer stick on the wikipedia syntax definition ?

I though that the `GlobIterator` in PHP is not able to handle the braces but it must be checked.

For the moment the implementation is only using the `glob` function but detect if it's required or not...