@Jir4 2016-02-19
Can you run some bench ? It can be usefull to compare with and without the `preg_match`