@shulard 2016-02-19
Yep of course, I'll submit the results here!