@shulard 2016-02-19
I've updated my bench repo here: https://github.com/shulard/hoa-file-glob-benchmark

Now I check with `glob`, with `Iterator\Glob`, with detection+`glob`, with detection+`Iterator\Glob`:

```
glob              ||||||||||                                      1677ms,  22.0%
iterator          ||||||||||                                      1723ms,  22.6%
globnodetect      |||||||||||                                     1800ms,  23.6%
iteratornodetect  ||||||||||                                      1667ms,  21.9%
```

Execution time is really similar with the given path patterns :

* `/Users/shulard/Sites/Clients`
* `/[u]sr/[lb]*/`

All the objects found 42083 files on my computer.
I ran the test multiple times, `Iterator\Glob` results are all consistant but `glob` can take about 100ms more on some calls... Maybe it's my computer, maybe it's the implementation...

I also confirmed that braces are not handled by the `GlobIterator` implementation in PHP (but it's not an official feature of glob...).