@shulard 2016-02-22
I think that we need to use the `GlobIterator` which is more consistent. Then the brace support is not really important in our case (I think...).