@shulard 2016-03-21
Hello !

What's your POV here ? Do you think we need to use the GlobIterator too ?
I can update the PR code with the final method during this week.