@Hywan 2016-03-23
I think `GlobIterator` is more consistent, but… why does it have less features? Today `Hoa\Iterator\Glob` extends `GlobIterator` from SPL (https://github.com/hoaproject/Iterator/blob/60bdefab8db17717871a11101dedec60572f95b8/Glob.php). You may want to ask on internals why `glob` and `GlobIterator` do not provide the same features. Based on these answers, we will see what to do next.

Thoughts?