@Hywan 2016-03-23
And yes, sometimes it can be very annoying to development inside Hoa because we have to check everything and every details, but this is why the quality is high :-p. Also, it often happens that we find inconsistency in PHP API or between PHP VM (remember [`Hoa\Iterator\RegularExpression`](https://github.com/hoaproject/Iterator/blob/60bdefab8db17717871a11101dedec60572f95b8/RegularExpression.php#L42-L46) —and `Hoa\Iterator\Recursive\RegularExpression`— with https://github.com/facebook/hhvm/issues/3909 and https://bugs.php.net/bug.php?id=68128). I wonder if we should communicate about this…