@shulard 2016-03-23
@Hywan, no problem about the triple-check :smile:. This process allow me to learn a lot about quality process and reviews, it's awesome to be part of that !

I'll make an email to php-src internal about the difference betweens the 2 globs.

About the discovered inconsistencies, I think Hoa should communicate about it! It'll help move forward the different projects and also show that Hoa is really making deeper research about it's code.