@shulard 2016-03-23
Thread created on php-general mailing list: http://news.php.net/php.general/325485