@shulard 2016-03-24
I got an answer from PHP ML : http://news.php.net/php.general/325487

It seems that `glob` is a porting of the linux's glob feature. The `GlobIterator` is more strict regarding the glob pattern implementation and we saw that there is no overhead when using an Iterator combination.

Because the specific `glob` features are enabled with flags, the `GlobIterator` is just equivalent to a `glob`  call with no flags.