@shulard 2016-03-25
Hello !
I've updated the implementation to use the `Iterator`.
Is the ML reply enough to validate that choice ?