@Hywan 2016-04-01
@shulard Good work! Very proud of you. We can go for `Hoa\Iterator\Glob`, it's fine.