@shulard 2016-04-01
Thanks a lot for the review, I've updated the code with your feedback.

The regex pattern is really simpler :smile:.